refactor: 타입, 함수 컴포넌트 분리
